[service] Export a separate product object
